fix(evalhub): address review findings before team review
- Move evalhub stub from test_adapter.py to conftest.py so tests work
  regardless of collection order or file selection
- Wrap MLflow trace enrichment in its own try/except to prevent flaky
  MLflow from invalidating successful query results
- Validate missing 'query' key in YAML entries with descriptive
  ValueError; widen adapter's load_queries catch to include ValueError
- Strip orphan difficulty/category fields from tool_use.yaml
- Add __post_init__ validation for positive timeout/latency values
- Broaden .gitignore REFACTORING.md pattern to **/REFACTORING.md
- Add TODO for asyncio.gather concurrent query execution
- Update README: validation docs, MLflow fault tolerance, sequential
  execution note, conftest.py stub explanation

Made-with: Cursor
