+++ .gitignore
@@ -18,4 +18,4 @@ agents/langflow/simple_tool_calling_agent/local/.ollama-enabled
 *.db
 CLAUDE.local.md
 .cursor
-tests/behavioral/evalhub_adapter/REFACTORING.md
+**/REFACTORING.md
